refactor: restructure component registry by moving PropsParser and Registry to a dedicated directory, updating imports accordingly
