+++ packages/@withstudiocms/component-registry/src/index.ts
@@ -1,8 +1,6 @@
 export * from './component-proxy/index.js';
 export * from './errors.js';
-export * from './handler.js';
-export * from './PropsParser.js';
-export * from './Registry.js';
+export * from './registry/index.js';
 export * from './transform-html.js';
 export * from './types.js';
 export * from './utils.js';
