+++ packages/@withstudiocms/component-registry/src/registry/index.ts
@@ -1,0 +1,3 @@
+export * from './handler.js';
+export * from './PropsParser.js';
+export * from './Registry.js';
